# swagger.yaml
openapi: 3.0.0
info:
  title: GOSHOP-ECOMMERCE-API
  contact: {}
  version: '1.0'
servers:
- url: https://goshop-vercel.vercel.app/api/v1
  variables: {}
paths:
  /products:
    get:
      tags:
      - Product
      summary: Get All Products
      operationId: GetAllProducts
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
    post:
      tags:
      - Product
      summary: Create Product
      operationId: CreateProduct
      parameters: []
      requestBody:
        content:
          multipart/form-data:
            encoding: {}
            schema:
              required:
              - name
              - oldPrice
              - price
              - description
              - content
              - category
              - isStocked
              - productNumber
              - isSale
              - isNew
              - image
              - imageGallery
              - filterItems
              - colors
              - shippingFees
              type: object
              properties:
                name:
                  type: string
                  example: Detox body Cream
                oldPrice:
                  type: number
                  format: double
                  example: 149.95
                price:
                  type: number
                  format: double
                  example: 99.95
                description:
                  type: string
                  example: Lorem ipsum dolor sit amet consectetur adipisicing elit. Modi beatae provident ipsum omnis dolor sapiente maiores reiciendis exercitationem earum deleniti, reprehenderit iste ipsa saepe. Consectetur non et excepturi molestias esse
                content:
                  type: string
                  example: Universal foundation masks skin imperfections and gives it a matte finish, remaining completely invisible on the face. The foundation is suitable for any skin type. Thanks to a unique combination of pigments, the foundation adapts perfectly to even the smallest features of skin tone, creating an even and natural tone
                category:
                  type: string
                  example: skin
                isStocked:
                  type: boolean
                  example: true
                productNumber:
                  type: string
                  example: IN1203
                isSale:
                  type: boolean
                  example: false
                isNew:
                  type: boolean
                  example: true
                image:
                  type: string
                  format: binary
                imageGallery:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: ''
                filterItems:
                  type: array
                  items:
                    type: string
                  description: ''
                colors:
                  type: array
                  items:
                    type: string
                  description: ''
                shippingFees:
                  type: number
                  format: double
                  example: 10.99
        required: false
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
  /products/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: The product ID
    get:
      tags:
      - Product
      summary: Get Product
      operationId: GetProduct
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
    patch:
      tags:
      - Product
      summary: Update Product
      operationId: UpdateProduct
      parameters: []
      requestBody:
        content:
          multipart/form-data:
            encoding: {}
            schema:
              required:
              - name
              - oldPrice
              - price
              - description
              - content
              - category
              - isStocked
              - productNumber
              - isSale
              - isNew
              - filterItems
              - colors
              - image
              - imageGallery
              type: object
              properties:
                name:
                  type: string
                  example: Detox body Cream
                oldPrice:
                  type: number
                  format: double
                  example: 149.95
                price:
                  type: number
                  format: double
                  example: 99.99
                description:
                  type: string
                  example: Lorem ipsum dolor sit amet consectetur adipisicing elit. Modi beatae provident ipsum omnis dolor sapiente maiores reiciendis exercitationem earum deleniti, reprehenderit iste ipsa saepe. Consectetur non et excepturi molestias esse
                content:
                  type: string
                  example: Universal foundation masks skin imperfections and gives it a matte finish, remaining completely invisible on the face. The foundation is suitable for any skin type. Thanks to a unique combination of pigments, the foundation adapts perfectly to even the smallest features of skin tone, creating an even and natural tone
                category:
                  type: string
                  example: skin
                isStocked:
                  type: boolean
                  example: true
                productNumber:
                  type: string
                  example: IN1203
                isSale:
                  type: boolean
                  example: false
                isNew:
                  type: boolean
                  example: false
                filterItems:
                  type: array
                  items:
                    type: string
                  description: ''
                colors:
                  type: array
                  items:
                    type: string
                  description: ''
                image:
                  type: string
                  format: binary
                imageGallery:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: ''
        required: false
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    delete:
      tags:
      - Product
      summary: Delete Product
      operationId: DeleteProduct
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
...
  /reviews:
    get:
      tags:
      - Reviews
      summary: Get All Reviews
      operationId: GetAllReviews
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
    post:
      tags:
      - Reviews
      summary: Create Review
      operationId: CreateReview
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/CreateReviewRequest'
              - example:
                  title: Good Product
                  rating: 4
                  comment: This is just what I need. thanks you for the best customer support and fast shipping
                  user: 64363fd7358253c2d0b03a69
                  product: 643b1d67896a6e693e7c8823
            example:
              title: Good Product
              rating: 4
              comment: This is just what I need. thanks you for the best customer support and fast shipping
              user: 64363fd7358253c2d0b03a69
              product: 643b1d67896a6e693e7c8823
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
...
  /orders/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: The order ID
    get:
      tags:
      - Orders
      summary: Get Single Order
      operationId: GetSingleOrder
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    patch:
      tags:
      - Orders
      summary: Update Order
      operationId: UpdateOrder
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
...
components:
  schemas:
...
    CreateReviewRequest:
      title: CreateReviewRequest
      required:
      - title
      - rating
      - comment
      - user
      - product
      type: object
      properties:
        title:
          type: string
        rating:
          type: integer
          format: int32
        comment:
          type: string
        user:
          type: string
        product:
          type: string
      example:
        title: Good Product
        rating: 4
        comment: This is just what I need. thanks you for the best customer support and fast shipping
        user: 64363fd7358253c2d0b03a69
        product: 643b1d67896a6e693e7c8823
  securitySchemes:
    httpBearer:
      type: http
      scheme: bearer
security:
- httpBearer: []
...